Trekker ut MottakerInfo fra JournalførVedtaksbrevDTO til en egen fil under brev/mottaker
