+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/brev/DokumentService.kt
@@ -18,14 +18,14 @@ import no.nav.familie.ba.sak.kjerne.brev.domene.ManuellBrevmottaker
 import no.nav.familie.ba.sak.kjerne.brev.domene.ManueltBrevRequest
 import no.nav.familie.ba.sak.kjerne.brev.domene.erTilInstitusjon
 import no.nav.familie.ba.sak.kjerne.brev.mottaker.BrevmottakerService
+import no.nav.familie.ba.sak.kjerne.brev.mottaker.Bruker
+import no.nav.familie.ba.sak.kjerne.brev.mottaker.FullmektigEllerVerge
+import no.nav.familie.ba.sak.kjerne.brev.mottaker.Institusjon
+import no.nav.familie.ba.sak.kjerne.brev.mottaker.MottakerInfo
+import no.nav.familie.ba.sak.kjerne.brev.mottaker.tilAvsenderMottaker
 import no.nav.familie.ba.sak.kjerne.fagsak.Fagsak
 import no.nav.familie.ba.sak.kjerne.fagsak.FagsakRepository
 import no.nav.familie.ba.sak.kjerne.steg.BehandlerRolle
-import no.nav.familie.ba.sak.kjerne.steg.domene.Bruker
-import no.nav.familie.ba.sak.kjerne.steg.domene.FullmektigEllerVerge
-import no.nav.familie.ba.sak.kjerne.steg.domene.Institusjon
-import no.nav.familie.ba.sak.kjerne.steg.domene.MottakerInfo
-import no.nav.familie.ba.sak.kjerne.steg.domene.tilAvsenderMottaker
 import no.nav.familie.ba.sak.kjerne.steg.grunnlagForNyBehandling.VilkårsvurderingForNyBehandlingService
 import no.nav.familie.ba.sak.kjerne.vedtak.Vedtak
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingService
