+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/steg/JournalførVedtaksbrev.kt
@@ -13,16 +13,16 @@ import no.nav.familie.ba.sak.kjerne.brev.DokumentService.Companion.genererEkster
 import no.nav.familie.ba.sak.kjerne.brev.domene.ManuellBrevmottaker
 import no.nav.familie.ba.sak.kjerne.brev.hentOverstyrtDokumenttittel
 import no.nav.familie.ba.sak.kjerne.brev.mottaker.BrevmottakerService
+import no.nav.familie.ba.sak.kjerne.brev.mottaker.Bruker
+import no.nav.familie.ba.sak.kjerne.brev.mottaker.BrukerMedUtenlandskAdresse
+import no.nav.familie.ba.sak.kjerne.brev.mottaker.Dødsbo
+import no.nav.familie.ba.sak.kjerne.brev.mottaker.FullmektigEllerVerge
+import no.nav.familie.ba.sak.kjerne.brev.mottaker.Institusjon
+import no.nav.familie.ba.sak.kjerne.brev.mottaker.MottakerInfo
+import no.nav.familie.ba.sak.kjerne.brev.mottaker.tilAvsenderMottaker
 import no.nav.familie.ba.sak.kjerne.fagsak.FagsakRepository
 import no.nav.familie.ba.sak.kjerne.fagsak.FagsakType
-import no.nav.familie.ba.sak.kjerne.steg.domene.Bruker
-import no.nav.familie.ba.sak.kjerne.steg.domene.BrukerMedUtenlandskAdresse
-import no.nav.familie.ba.sak.kjerne.steg.domene.Dødsbo
-import no.nav.familie.ba.sak.kjerne.steg.domene.FullmektigEllerVerge
-import no.nav.familie.ba.sak.kjerne.steg.domene.Institusjon
 import no.nav.familie.ba.sak.kjerne.steg.domene.JournalførVedtaksbrevDTO
-import no.nav.familie.ba.sak.kjerne.steg.domene.MottakerInfo
-import no.nav.familie.ba.sak.kjerne.steg.domene.tilAvsenderMottaker
 import no.nav.familie.ba.sak.kjerne.vedtak.Vedtak
 import no.nav.familie.ba.sak.kjerne.vedtak.VedtakService
 import no.nav.familie.ba.sak.task.DistribuerDokumentDTO
